learnGoWithTests/helloworld/hello_test.go
package main

import (
	"testing"
)

// Testing is writen into the language; no external packages necessary
// Its best to be writing the tests first in many circumstances
func TestHello (t *testing.T) {
	t.Run("saying hello to people", func(t *testing.T) {
		got := Hello("Chris", "")
		want := "Hello, Chris"
		assertCorrectMessage(t, got, want)
	})

	t.Run("say 'Hello, World' when an empty string is supplied", func(t *testing.T) {
		got := Hello("", "")
		want := "Hello, World"
		assertCorrectMessage(t, got, want)
	})

	t.Run("in Spanish", func(t *testing.T) {
		got := Hello("Elodie", "Spanish")
		want := "Hola, Elodie"
		assertCorrectMessage(t, got, want)
	})

	t.Run("in French", func(t *testing.T) {
		got := Hello("Pierre", "French")
		want := "Bonjour, Pierre"
		assertCorrectMessage(t, got, want)
	})
}

func assertCorrectMessage(t testing.TB, got, want string) {
	t.Helper() // This will let Go know that this is a helper function
	// it will redirect the programmer to the correct line above rather than the error statement below

	if got != want {
		t.Errorf("got %q want %q", got, want)
	}
}